Refactor hooks and imports in `useLanguage.js` and related components; - Update import statements and useCallback usage - Add schema validation logic using useMemo and useState + Use destructuring to simplify imports from 'react' + Import axiosPrivate from a separate file for better organization
Changes in `AdminSettings.jsx`:
- Update import statement for 'useLanguage' hook
- Change React hooks to functional component hooks (useState, useMemo, useState)

Changes in `LanguageForm.jsx`:
- No changes related to the provided diff snippet.
