+++ src/pages/Language.jsx
@@ -1,4 +1,5 @@
-import React, { useCallback, useEffect, useState } from "react";
+import { useCallback, useEffect, useState } from "react";
+
 import PropTypes from "prop-types";
 
 import { axiosPrivate } from "~/_api";
